Update docker-build and after_script in .gitlab-ci.yml
Updated the `.gitlab-ci.yml` file with significant changes in the `docker-build` and `after_script` sections. In `docker-build`, added commands to create a new builder instance `mybuilder`, pause the script for 2 seconds, and build and push the Docker image to the `serpensin/dbdstats:latest` repository. In `after_script`, added a command to remove the `mybuilder` instance and removed commands that paused the script for 10 seconds and pruned all unused Docker volumes. The `timeout` value and `workflow` section remain unchanged.