Update app version and modify emoji codes in embed descriptions
The Dockerfile and `bot_version` variable in `main.py` have been updated to reflect the new application version, changing from 1.14.3 to 1.14.4. Additionally, the descriptions of the `embed` objects in the `Info` class in `main.py` have been updated, with specific emoji codes in the descriptions for the current event, upcoming event, and no events being replaced with new ones.