Updated application version and improved function comparisons
Updated the application version from "1.14.6" to "1.14.7" in the Dockerfile and main.py. Enhanced the `find_key_by_name` function in the `Functions` class to make the comparison of `key_name` and `input_name` case-insensitive. Commented out the replacement of spaces in the `perk` variable in the `SendInfo` class, allowing the `perk` string to retain spaces.